Updated with minimal information
updated info:

Handlers are javascript scripts that handle a game. There is one handler for each game. A handler tells Splitscreen.me how and what to open. 
Basically, if you want to splitscreen a game you need a handler to tell Splitscreen.me what and how to splitscreen it. 

For info on how to make a handler for yourself, please see [create your handler](https://www.splitscreen.me/docs/create-handlers).